seek修改
在独立线程seek（偶尔崩溃，死锁，待修复）
添加seekbar
总时长不对（已改）

+++ AVPlayer/FrameData.h
@@ -35,7 +35,7 @@ public:
 	mutable int height_;
 	mutable int size_;
 	mutable BYTE* data_;
-	mutable double tm_;
+	mutable int64_t tm_;
 };
